pyproject.toml
[tool.poetry]
name = "chocobot"
version = "0.1.0"
description = ""
authors = ["Cassidy Dingenskirchen <admin@15318.de>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chocobot = "chocobot.__main__:main"
